Add waiting room and add-session flow
Introduces <WaitingRoom> and its subcomponents to list all current game
sessions. Currently hard codes 'chat-room' and 'tic-tac-toe' as the two
available games (to be implemented later).

The component listens for changes to the session metadata database to
populate the waiting room cards. Shows a spinner until loading the
initial list. The add-session flow pushes new session metadata directly
to the Firebase Realtime Database.

The join-game flow is simply a link to /<session-type>/<session-id>.
That route will load the (to be implemented) game component.